Move quantum control tests into a separate file
